+++ src/providers/local.ts
@@ -5,7 +5,7 @@ import { LocalProviderCore } from './local-core'
 import { SqliteLocalStore } from './sqlite-local-store'
 
 export class LocalProvider extends LocalProviderCore {
-  constructor(db: Database, dbPath = getDbPath()) {
-    super(new SqliteLocalStore(db, dbPath))
+  constructor(db: Database, dbPath = getDbPath(), defaultTaskColumn?: string) {
+    super(new SqliteLocalStore(db, dbPath, defaultTaskColumn))
   }
 }